Added triangulate() function.
