+++ vispy/geometry/__init__.py
@@ -8,13 +8,14 @@ This module implements classes and methods for handling geometric data.
 
 from __future__ import division
 
-__all__ = ['MeshData', 'PolygonData', 'Rect', 'Triangulation', 'create_arrow',
-           'create_cone', 'create_cube', 'create_cylinder', 'create_sphere']
+__all__ = ['MeshData', 'PolygonData', 'Rect', 'Triangulation', 'triangulate',
+           'create_arrow', 'create_cone', 'create_cube', 'create_cylinder',
+           'create_sphere']
 
 from .polygon import PolygonData  # noqa
 from .meshdata import MeshData  # noqa
 from .rect import Rect  # noqa
-from .triangulation import Triangulation  # noqa
+from .triangulation import Triangulation, triangulate  # noqa
 from .torusknot import TorusKnot  # noqa
 from .calculations import _calculate_normals, _fast_cross_3d  # noqa
 from .generation import create_arrow, create_cone, create_cube, \
